+++ DutWrapper/Account.cs
@@ -2,12 +2,15 @@
 using AngleSharp.Dom;
 using DutWrapper.Model;
 using DutWrapper.Model.Account;
+using DutWrapper.Model.Enums;
+using DutWrapper.Utils;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
+using System.Xml.Linq;
 
 namespace DutWrapper
 {
@@ -16,24 +19,21 @@ namespace DutWrapper
         private static string BASE_ADDRESS = "http://sv.dut.udn.vn";
         private static string __VIEWSTATE = "/wEPDwUKMTY2NjQ1OTEyNA8WAh4TVmFsaWRhdGVSZXF1ZXN0TW9kZQIBFgJmD2QWAgIFDxYCHglpbm5lcmh0bWwF4i08dWwgaWQ9J21lbnUnIHN0eWxlPSd3aWR0aDogMTI4MHB4OyBtYXJnaW46IDAgYXV0bzsgJz48bGk+PGEgSUQ9ICdsUGFIT01FJyBzdHlsZSA9J3dpZHRoOjY1cHgnIGhyZWY9J0RlZmF1bHQuYXNweCc+VHJhbmcgY2jhu6c8L2E+PGxpPjxhIElEPSAnbFBhQ1REVCcgc3R5bGUgPSd3aWR0aDo4NXB4JyBocmVmPScnPkNoxrDGoW5nIHRyw6xuaDwvYT48dWwgY2xhc3M9J3N1Ym1lbnUnPjxsaT48YSBJRCA9J2xDb0NURFRDMicgICBzdHlsZSA9J3dpZHRoOjE0MHB4JyBocmVmPSdHX0xpc3RDVERULmFzcHgnPkNoxrDGoW5nIHRyw6xuaCDEkcOgbyB04bqhbzwvYT48L2xpPjxsaT48YSBJRCA9J2xDb0NURFRDMScgICBzdHlsZSA9J3dpZHRoOjE0MHB4JyBocmVmPSdHX0xpc3RIb2NQaGFuLmFzcHgnPkjhu41jIHBo4bqnbjwvYT48L2xpPjxsaT48YSBJRCA9J2xDb0NURFRDMycgICBzdHlsZSA9J3dpZHRoOjIwMHB4JyBocmVmPSdHX0xpc3RDVERUQW5oLmFzcHgnPlByb2dyYW08L2E+PC9saT48L3VsPjwvbGk+PGxpPjxhIElEPSAnbFBhS0hEVCcgc3R5bGUgPSd3aWR0aDo2MHB4JyBocmVmPScnPkvhur8gaG/huqFjaDwvYT48dWwgY2xhc3M9J3N1Ym1lbnUnPjxsaT48YSBJRCA9J2xDb0tIRFRDMScgICBzdHlsZSA9J3dpZHRoOjIwMHB4JyBocmVmPSdodHRwczovLzFkcnYubXMvYi9zIUF0d0tsRFo2VnFidG5RY2JqUVFwS05rbUswX2g/ZT1uQ2I3eVAnPkvhur8gaG/huqFjaCDEkcOgbyB04bqhbyBuxINtIGjhu41jPC9hPjwvbGk+PGxpPjxhIElEID0nbENvS0hEVEMyJyAgIHN0eWxlID0nd2lkdGg6MjAwcHgnIGhyZWY9J2h0dHA6Ly9kazQuZHV0LnVkbi52bic+xJDEg25nIGvDvSBo4buNYzwvYT48L2xpPjxsaT48YSBJRCA9J2xDb0tIRFRDMycgICBzdHlsZSA9J3dpZHRoOjIwMHB4JyBocmVmPSdodHRwOi8vZGs0LmR1dC51ZG4udm4vR19Mb3BIb2NQaGFuLmFzcHgnPkzhu5twIGjhu41jIHBo4bqnbiAtIMSRYW5nIMSRxINuZyBrw708L2E+PC9saT48bGk+PGEgSUQgPSdsQ29LSERUQzQnICAgc3R5bGUgPSd3aWR0aDoyMDBweCcgaHJlZj0nR19Mb3BIb2NQaGFuLmFzcHgnPkzhu5twIGjhu41jIHBo4bqnbiAtIGNow61uaCB0aOG7qWM8L2E+PC9saT48bGk+PGEgSUQgPSdsQ29LSERUQzUnICAgc3R5bGUgPSd3aWR0aDoyMDBweCcgaHJlZj0naHR0cDovL2RrNC5kdXQudWRuLnZuL0dfREt5Tmh1Q2F1LmFzcHgnPkto4bqjbyBzw6F0IG5odSBj4bqndSBt4bufIHRow6ptIGzhu5twPC9hPjwvbGk+PGxpPjxhIElEID0nbENvS0hEVEM2JyAgIHN0eWxlID0nd2lkdGg6MjAwcHgnIGhyZWY9J2h0dHA6Ly9jYi5kdXQudWRuLnZuL1BhZ2VMaWNoVGhpS0guYXNweCc+VGhpIGN14buRaSBr4buzIGzhu5twIGjhu41jIHBo4bqnbjwvYT48L2xpPjxsaT48YSBJRCA9J2xDb0tIRFRDNycgICBzdHlsZSA9J3dpZHRoOjIwMHB4JyBocmVmPSdHX0RLVGhpTk4uYXNweCc+VGhpIFRp4bq/bmcgQW5oIMSR4buLbmgga+G7sywgxJHhuqd1IHJhPC9hPjwvbGk+PGxpPjxhIElEID0nbENvS0hEVEM4JyAgIHN0eWxlID0nd2lkdGg6MjAwcHgnIGhyZWY9J0dfTGlzdExpY2hTSC5hc3B4Jz5TaW5oIGhv4bqhdCBs4bubcCDEkeG7i25oIGvhu7M8L2E+PC9saT48bGk+PGEgSUQgPSdsQ29LSERUQzknICAgc3R5bGUgPSd3aWR0aDoyMDBweCcgaHJlZj0naHR0cDovL2ZiLmR1dC51ZG4udm4nPkto4bqjbyBzw6F0IMO9IGtp4bq/biBzaW5oIHZpw6puPC9hPjwvbGk+PGxpPjxhIElEID0nbENvS0hEVEM5JyAgIHN0eWxlID0nd2lkdGg6MjAwcHgnIGhyZWY9J0dfREtQVkNELmFzcHgnPkhv4bqhdCDEkeG7mW5nIHBo4bulYyB24bulIGPhu5luZyDEkeG7k25nPC9hPjwvbGk+PC91bD48L2xpPjxsaT48YSBJRD0gJ2xQYVRSQUNVVScgc3R5bGUgPSd3aWR0aDo3MHB4JyBocmVmPScnPkRhbmggc8OhY2g8L2E+PHVsIGNsYXNzPSdzdWJtZW51Jz48bGk+PGEgSUQgPSdsQ29UUkFDVVUwMScgICBzdHlsZSA9J3dpZHRoOjI0MHB4JyBocmVmPSdHX0xpc3ROZ3VuZ0hvYy5hc3B4Jz5TaW5oIHZpw6puIG5n4burbmcgaOG7jWM8L2E+PC9saT48bGk+PGEgSUQgPSdsQ29UUkFDVVUwMycgICBzdHlsZSA9J3dpZHRoOjI0MHB4JyBocmVmPSdHX0xpc3RMb3AuYXNweCc+U2luaCB2acOqbiDEkWFuZyBo4buNYyAtIGzhu5twPC9hPjwvbGk+PGxpPjxhIElEID0nbENvVFJBQ1VVMDQnICAgc3R5bGUgPSd3aWR0aDoyNDBweCcgaHJlZj0nR19MaXN0Q0NDTlRULmFzcHgnPlNpbmggdmnDqm4gY8OzIGNo4bupbmcgY2jhu4kgQ05UVDwvYT48L2xpPjxsaT48YSBJRCA9J2xDb1RSQUNVVTA1JyAgIHN0eWxlID0nd2lkdGg6MjQwcHgnIGhyZWY9J0dfTGlzdENDTk4uYXNweCc+U2luaCB2acOqbiBjw7MgY2jhu6luZyBjaOG7iSBuZ2/huqFpIG5n4buvPC9hPjwvbGk+PGxpPjxhIElEID0nbENvVFJBQ1VVMDYnICAgc3R5bGUgPSd3aWR0aDoyNDBweCcgaHJlZj0naHR0cDovL2Rhb3Rhby5kdXQudWRuLnZuL1NWL0dfS1F1YUFuaFZhbi5hc3B4Jz5TaW5oIHZpw6puIHRoaSBUaeG6v25nIEFuaCDEkeG7i25oIGvhu7M8L2E+PC9saT48bGk+PGEgSUQgPSdsQ29UUkFDVVUwNycgICBzdHlsZSA9J3dpZHRoOjI0MHB4JyBocmVmPSdHX0xpc3REb0FuVE4uYXNweCc+U2luaCB2acOqbiBsw6BtIMSQ4buTIMOhbiB04buRdCBuZ2hp4buHcDwvYT48L2xpPjxsaT48YSBJRCA9J2xDb1RSQUNVVTA4JyAgIHN0eWxlID0nd2lkdGg6MjQwcHgnIGhyZWY9J0dfTGlzdEhvYW5Ib2NQaGkuYXNweCc+U2luaCB2acOqbiDEkcaw4bujYyBob8OjbiDEkcOzbmcgaOG7jWMgcGjDrTwvYT48L2xpPjxsaT48YSBJRCA9J2xDb1RSQUNVVTE2JyAgIHN0eWxlID0nd2lkdGg6MjQwcHgnIGhyZWY9J0dfTGlzdEhvYW5fVGhpQlMuYXNweCc+U2luaCB2acOqbiDEkcaw4bujYyBob8OjbiB0aGksIHRoaSBi4buVIHN1bmc8L2E+PC9saT48bGk+PGEgSUQgPSdsQ29UUkFDVVUwOScgICBzdHlsZSA9J3dpZHRoOjI0MHB4JyBocmVmPSdHX0xpc3RIb2NMYWkuYXNweCc+U2luaCB2acOqbiBk4buxIHR1eeG7g24gdsOgbyBo4buNYyBs4bqhaTwvYT48L2xpPjxsaT48YSBJRCA9J2xDb1RSQUNVVTEwJyAgIHN0eWxlID0nd2lkdGg6MjQwcHgnIGhyZWY9J0dfTGlzdEt5THVhdC5hc3B4Jz5TaW5oIHZpw6puIGLhu4sga+G7tyBsdeG6rXQ8L2E+PC9saT48bGk+PGEgSUQgPSdsQ29UUkFDVVUxMScgICBzdHlsZSA9J3dpZHRoOjI0MHB4JyBocmVmPSdHX0xpc3RCaUh1eUhQLmFzcHgnPlNpbmggdmnDqm4gYuG7iyBo4buneSBo4buNYyBwaOG6p248L2E+PC9saT48bGk+PGEgSUQgPSdsQ29UUkFDVVUxMicgICBzdHlsZSA9J3dpZHRoOjI0MHB4JyBocmVmPSdHX0xpc3RMb2NrV2ViLmFzcHgnPlNpbmggdmnDqm4gYuG7iyBraMOzYSB3ZWJzaXRlPC9hPjwvbGk+PGxpPjxhIElEID0nbENvVFJBQ1VVMTMnICAgc3R5bGUgPSd3aWR0aDoyNDBweCcgaHJlZj0nR19MaXN0TG9ja1dlYlRhbS5hc3B4Jz5TaW5oIHZpw6puIGLhu4sgdOG6oW0ga2jDs2Egd2Vic2l0ZTwvYT48L2xpPjxsaT48YSBJRCA9J2xDb1RSQUNVVTE0JyAgIHN0eWxlID0nd2lkdGg6MjQwcHgnIGhyZWY9J0dfTGlzdEhhbkNoZVRDLmFzcHgnPlNpbmggdmnDqm4gYuG7iyBo4bqhbiBjaOG6vyB0w61uIGNo4buJIMSRxINuZyBrw708L2E+PC9saT48bGk+PGEgSUQgPSdsQ29UUkFDVVUxNScgICBzdHlsZSA9J3dpZHRoOjI0MHB4JyBocmVmPSdHX0xpc3RDYW5oQmFvS1FIVC5hc3B4Jz5TaW5oIHZpw6puIGLhu4sgY+G6o25oIGLDoW8ga+G6v3QgcXXhuqMgaOG7jWMgdOG6rXA8L2E+PC9saT48L3VsPjwvbGk+PGxpPjxhIElEPSAnbFBhQ1VVU1YnIHN0eWxlID0nd2lkdGg6ODhweCcgaHJlZj0nJz5D4buxdSBzaW5oIHZpw6puPC9hPjx1bCBjbGFzcz0nc3VibWVudSc+PGxpPjxhIElEID0nbENvQ1VVU1YxJyAgIHN0eWxlID0nd2lkdGg6MTEwcHgnIGhyZWY9J0dfTGlzdFROZ2hpZXAuYXNweCc+xJDDoyB04buRdCBuZ2hp4buHcDwvYT48L2xpPjxsaT48YSBJRCA9J2xDb0NVVVNWMicgICBzdHlsZSA9J3dpZHRoOjExMHB4JyBocmVmPSdHX0xpc3RLaG9uZ1ROLmFzcHgnPktow7RuZyB04buRdCBuZ2hp4buHcDwvYT48L2xpPjwvdWw+PC9saT48bGk+PGEgSUQ9ICdsUGFDU1ZDJyBzdHlsZSA9J3dpZHRoOjE0NXB4JyBocmVmPScnPlBow7JuZyBo4buNYyAmIEjhu4cgdGjhu5FuZzwvYT48dWwgY2xhc3M9J3N1Ym1lbnUnPjxsaT48YSBJRCA9J2xDb0NTVkMwMScgICBzdHlsZSA9J3dpZHRoOjI0MHB4JyBocmVmPSdodHRwOi8vY2IuZHV0LnVkbi52bi9QYWdlQ05QaG9uZ0hvYy5hc3B4Jz5Uw6xuaCBow6xuaCBz4butIGThu6VuZyBwaMOybmcgaOG7jWM8L2E+PC9saT48bGk+PGEgSUQgPSdsQ29DU1ZDMDInICAgc3R5bGUgPSd3aWR0aDoyNDBweCcgaHJlZj0nR19MaXN0VGhCaUhvbmcuYXNweCc+VGjhu5FuZyBrw6ogYsOhbyB0aGnhur90IGLhu4sgcGjDsm5nIGjhu41jIGjhu49uZzwvYT48L2xpPjxsaT48YSBJRCA9J2xDb0NTVkMwOScgICBzdHlsZSA9J3dpZHRoOjI0MHB4JyBocmVmPSdHX1N5c1N0YXR1cy5hc3B4Jz5UcuG6oW5nIHRow6FpIGjhu4cgdGjhu5FuZyB0aMO0bmcgdGluIHNpbmggdmnDqm48L2E+PC9saT48L3VsPjwvbGk+PGxpPjxhIElEPSAnbFBhTElFTktFVCcgc3R5bGUgPSd3aWR0aDo1MHB4JyBocmVmPScnPkxpw6puIGvhur90PC9hPjx1bCBjbGFzcz0nc3VibWVudSc+PGxpPjxhIElEID0nbENvTElFTktFVDEnICAgc3R5bGUgPSd3aWR0aDo3MHB4JyBocmVmPSdodHRwOi8vbGliLmR1dC51ZG4udm4nPlRoxrAgdmnhu4duPC9hPjwvbGk+PGxpPjxhIElEID0nbENvTElFTktFVDInICAgc3R5bGUgPSd3aWR0aDo3MHB4JyBocmVmPSdodHRwOi8vbG1zMS5kdXQudWRuLnZuJz5EVVQtTE1TPC9hPjwvbGk+PC91bD48L2xpPjxsaT48YSBJRD0gJ2xQYUhFTFAnIHN0eWxlID0nd2lkdGg6NDVweCcgaHJlZj0nJz5I4buXIHRy4bujPC9hPjx1bCBjbGFzcz0nc3VibWVudSc+PGxpPjxhIElEID0nbENvSEVMUDEnICAgc3R5bGUgPSd3aWR0aDoyMTBweCcgaHJlZj0naHR0cDovL2ZyLmR1dC51ZG4udm4nPkPhu5VuZyBo4buXIHRy4bujIHRow7RuZyB0aW4gdHLhu7FjIHR1eeG6v248L2E+PC9saT48bGk+PGEgSUQgPSdsQ29IRUxQMicgICBzdHlsZSA9J3dpZHRoOjIxMHB4JyBocmVmPSdodHRwczovL2RyaXZlLmdvb2dsZS5jb20vZmlsZS9kLzFaMHFsYmhLYVNHbXpFWkpEMnVCNGVVV2VlSGFROUhIbC92aWV3Jz5IxrDhu5tuZyBk4bqrbiDEkMSDbmcga8O9IGjhu41jPC9hPjwvbGk+PGxpPjxhIElEID0nbENvSEVMUDMnICAgc3R5bGUgPSd3aWR0aDoyMTBweCcgaHJlZj0naHR0cDovL2Rhb3Rhby5kdXQudWRuLnZuL2Rvd25sb2FkMi9FbWFpbF9HdWlkZS5wZGYnPkjGsOG7m25nIGThuqtuIFPhu60gZOG7pW5nIEVtYWlsIERVVDwvYT48L2xpPjxsaT48YSBJRCA9J2xDb0hFTFA3JyAgIHN0eWxlID0nd2lkdGg6MjEwcHgnIGhyZWY9J2h0dHBzOi8vMWRydi5tcy91L3MhQXR3S2xEWjZWcWJ0bzEwYmhIYzBLN3NleU5Hcj9lYUNUYjh4Jz5WxINuIGLhuqNuIFF1eSDEkeG7i25oIGPhu6dhIFRyxrDhu51uZzwvYT48L2xpPjxsaT48YSBJRCA9J2xDb0hFTFA4JyAgIHN0eWxlID0nd2lkdGg6MjEwcHgnIGhyZWY9J2h0dHBzOi8vdGlueXVybC5jb20veTRrZGozc3AnPkJp4buDdSBt4bqrdSB0aMaw4budbmcgZMO5bmc8L2E+PC9saT48L3VsPjwvbGk+PGxpPjxhIGlkID0nbGlua0RhbmdOaGFwJyBocmVmPSdQYWdlRGFuZ05oYXAuYXNweCcgc3R5bGUgPSd3aWR0aDo4MHB4Oyc+IMSQxINuZyBuaOG6rXAgPC9hPjwvbGk+PGxpPjxkaXYgY2xhc3M9J0xvZ2luRnJhbWUnPjxkaXYgc3R5bGUgPSdtaW4td2lkdGg6IDEwMHB4Oyc+PC9kaXY+PC9kaXY+PC9saT48L3VsPmRkFSwwNgHSdZ2bG7X5MK3ePxjwI3ZrE7W2esgf8K/1Yqk=";
 
-        private static FormUrlEncodedContent CreateLoginParameters(string username, string password)
-        {
-            Dictionary<string, string> dict = new Dictionary<string, string>
-            {
-                { "__VIEWSTATE", __VIEWSTATE },
-                { "__VIEWSTATEGENERATOR", "20CC0D2F" },
-                { "_ctl0:MainContent:DN_txtAcc", username },
-                { "_ctl0:MainContent:DN_txtPass", password },
-                { "_ctl0:MainContent:QLTH_btnLogin", "Đăng+nhập" }
-            };
-            return new FormUrlEncodedContent(dict);
-        }
-
-        public static string? GenerateNewSessionId()
+        private static HttpClient CreateDefaultHttpClient(string? sessionId = null)
         {
             HttpClient client = new HttpClient();
             client.BaseAddress = new Uri(BASE_ADDRESS);
-            HttpResponseMessage response = client.GetAsync($"/PageDangNhap.aspx").Result;
+            if (sessionId != null)
+            {
+                client.DefaultRequestHeaders.Add("Cookie", $"ASP.NET_SessionId={sessionId};");
+            }
+            return client;
+        }
+
+        public static async Task<string?> GenerateNewSessionId()
+        {
+            var client = CreateDefaultHttpClient();
+            HttpResponseMessage response = await client.GetAsync($"/PageDangNhap.aspx");
 
             string[] cookieHeaderList = new string[2] { "Set-Cookie", "Cookie" };
             foreach (string cookieHeader in cookieHeaderList)
@@ -61,18 +61,27 @@ namespace DutWrapper
 
         public static async Task<LoginStatus> Login(string sessionId, string username, string password)
         {
-            HttpClient client = new HttpClient();
-            client.BaseAddress = new Uri(BASE_ADDRESS);
-            client.DefaultRequestHeaders.Add("Cookie", $"ASP.NET_SessionId={sessionId};");
+            FormUrlEncodedContent CreateLoginParameters(string username, string password)
+            {
+                Dictionary<string, string> dict = new Dictionary<string, string>
+                {
+                    { "__VIEWSTATE", __VIEWSTATE },
+                    { "__VIEWSTATEGENERATOR", "20CC0D2F" },
+                    { "_ctl0:MainContent:DN_txtAcc", username },
+                    { "_ctl0:MainContent:DN_txtPass", password },
+                    { "_ctl0:MainContent:QLTH_btnLogin", "Đăng+nhập" }
+                };
+                return new FormUrlEncodedContent(dict);
+            }
+
+            var client = CreateDefaultHttpClient(sessionId);
             await client.PostAsync($"/PageDangNhap.aspx", CreateLoginParameters(username, password));
             return await IsLoggedIn(sessionId);
         }
 
         public static async Task<LoginStatus> IsLoggedIn(string sessionId)
         {
-            HttpClient client = new HttpClient();
-            client.BaseAddress = new Uri(BASE_ADDRESS);
-            client.DefaultRequestHeaders.Add("Cookie", $"ASP.NET_SessionId={ sessionId };");
+            var client = CreateDefaultHttpClient(sessionId);
             HttpResponseMessage response = await client.GetAsync($"/WebAjax/evLopHP_Load.aspx?E=TTKBLoad&Code=2120");
             if (response.IsSuccessStatusCode)
                 return LoginStatus.LoggedIn;
@@ -82,9 +91,7 @@ namespace DutWrapper
 
         public static async Task Logout(string sessionId)
         {
-            HttpClient client = new HttpClient();
-            client.BaseAddress = new Uri(BASE_ADDRESS);
-            client.DefaultRequestHeaders.Add("Cookie", $"ASP.NET_SessionId={sessionId};");
+            var client = CreateDefaultHttpClient(sessionId);
             await client.GetAsync($"/PageLogout.aspx");
         }
 
@@ -95,13 +102,11 @@ namespace DutWrapper
 
             List<SubjectSchedule>? result = new List<SubjectSchedule>();
 
-            HttpClient client = new HttpClient();
-            client.BaseAddress = new Uri(BASE_ADDRESS);
-            client.DefaultRequestHeaders.Add("Cookie", $"ASP.NET_SessionId={sessionId};");
+            var client = CreateDefaultHttpClient(sessionId);
             HttpResponseMessage response = await client.GetAsync($"/WebAjax/evLopHP_Load.aspx?E=TTKBLoad&Code={year}{(semester <= 2 ? semester : 2)}{(semester == 3 ? 1 : 0)}");
 
             if (!response.IsSuccessStatusCode)
-                throw new Exception(String.Format("The request has return code {0}.", response.StatusCode));
+                throw new Exception(string.Format("The request has return code {0}.", response.StatusCode));
 
             var config = Configuration.Default;
             var context = BrowsingContext.New(config);
@@ -126,8 +131,24 @@ namespace DutWrapper
                             Credit = ConvertTo<float>(cellCollection[3].TextContent),
                             IsHighQuality = cellCollection[5].ClassList.Contains("GridCheck"),
                             Lecturer = cellCollection[6].TextContent,
-                            ScheduleStudy = cellCollection[7].TextContent,
-                            Weeks = cellCollection[8].TextContent,
+                            ScheduleStudy = cellCollection[7].TextContent.Split("; ").Select(item1 =>
+                            {
+                                return new ScheduleItem(
+                                    item1.Split(",")[0].StartsWith("Thứ ") ? Convert.ToInt32(item1.Split(",")[0].Remove(0, 4)) : 1,
+                                    new Range(
+                                        Convert.ToInt32(item1.Split(",")[1].Split("-")[0]),
+                                        Convert.ToInt32(item1.Split(",")[1].Split("-")[1])
+                                        ),
+                                    item1.Split(",")[2]
+                                    );
+                            }).ToList(),
+                            Weeks = cellCollection[8].TextContent.Split(";").Select(item1 =>
+                            {
+                                return new Range(
+                                        Convert.ToInt32(item1.Split("-")[0]),
+                                        Convert.ToInt32(item1.Split("-")[1])
+                                    );
+                            }).ToList(),
                             PointFomula = cellCollection[10].TextContent
                         };
 
@@ -212,9 +233,7 @@ namespace DutWrapper
 
             List<SubjectFee>? result = new List<SubjectFee>();
 
-            HttpClient client = new HttpClient();
-            client.BaseAddress = new Uri(BASE_ADDRESS);
-            client.DefaultRequestHeaders.Add("Cookie", $"ASP.NET_SessionId={sessionId};");
+            var client = CreateDefaultHttpClient(sessionId);
             HttpResponseMessage response = await client.GetAsync($"/WebAjax/evLopHP_Load.aspx?E=THPhiLoad&Code={year}{(semester <= 2 ? semester : 2)}{(semester == 3 ? 1 : 0)}");
 
             if (!response.IsSuccessStatusCode)
@@ -241,7 +260,7 @@ namespace DutWrapper
                         item.Credit = ConvertTo<float>(cellCollection[3].TextContent);
                         item.IsHighQuality = cellCollection[4].ClassList.Contains("GridCheck");
                         item.Price = ConvertTo<double>(cellCollection[5].TextContent.Replace(",", null));
-                        item.Debt = cellCollection[6].ClassList.Contains("GridCheck");
+                        item.IsDebt = cellCollection[6].ClassList.Contains("GridCheck");
                         item.IsReStudy = cellCollection[7].ClassList.Contains("GridCheck");
                         item.VerifiedPaymentAt = cellCollection[8].TextContent;
 
@@ -259,10 +278,15 @@ namespace DutWrapper
 
         public static async Task<AccountInformation?> GetAccountInformation(string sessionId)
         {
-            string? GetIDFromTitleBar(string stringTitleBar)
+            string? GetIDFromTitleBar(string? stringTitleBar)
             {
                 try
                 {
+                    if (stringTitleBar == null)
+                    {
+                        throw new Exception("stringTitleBar is null here!");
+                    }
+
                     int openCase = stringTitleBar.IndexOf("(");
                     int closeCase = stringTitleBar.IndexOf(")");
                     int stringLength = (stringTitleBar.IndexOf(")")) - (stringTitleBar.IndexOf("(") + 1);
@@ -278,9 +302,7 @@ namespace DutWrapper
 
             try
             {
-                HttpClient client = new HttpClient();
-                client.BaseAddress = new Uri(BASE_ADDRESS);
-                client.DefaultRequestHeaders.Add("Cookie", $"ASP.NET_SessionId={sessionId};");
+                var client = CreateDefaultHttpClient(sessionId);
                 HttpResponseMessage response = await client.GetAsync($"/PageCaNhan.aspx");
 
                 if (!response.IsSuccessStatusCode)
@@ -290,14 +312,14 @@ namespace DutWrapper
                 var context = BrowsingContext.New(config);
                 var document = await context.OpenAsync(req => req.Content(response.Content.ReadAsStringAsync().Result));
 
-                accInfo.ID = GetIDFromTitleBar(document.GetElementById("Main_lblHoTen").TextContent);
+                accInfo.ID = GetIDFromTitleBar(document.GetElementById("Main_lblHoTen").GetTextContent());
                 accInfo.Name = document.GetElementById("CN_txtHoTen").GetValue();
-                accInfo.DateOfBirth = DateTime.ParseExact(document.GetElementById("CN_txtNgaySinh").GetValue(), "dd/MM/yyyy", CultureInfo.InvariantCulture);
-                accInfo.DatePlace = document.GetElementById("CN_cboNoiSinh").GetSelectedOptionOnSelectTag().TextContent;
-                accInfo.Ethnicity = document.GetElementById("CN_cboDanToc").GetSelectedOptionOnSelectTag().TextContent;
-                accInfo.Nationality = document.GetElementById("CN_cboQuocTich").GetSelectedOptionOnSelectTag().TextContent;
-                accInfo.Religion = document.GetElementById("CN_cboTonGiao").GetSelectedOptionOnSelectTag().TextContent;
-                switch (document.GetElementById("CN_txtGioiTinh").GetValue().ToLower())
+                accInfo.DateOfBirth = document.GetElementById("CN_txtNgaySinh").ConvertToDateTime();
+                accInfo.DatePlace = document.GetElementById("CN_cboNoiSinh").GetSelectedOptionOnSelectTag().GetTextContent();
+                accInfo.Ethnicity = document.GetElementById("CN_cboDanToc").GetSelectedOptionOnSelectTag().GetTextContent();
+                accInfo.Nationality = document.GetElementById("CN_cboQuocTich").GetSelectedOptionOnSelectTag().GetTextContent();
+                accInfo.Religion = document.GetElementById("CN_cboTonGiao").GetSelectedOptionOnSelectTag().GetTextContent();
+                switch (document.GetElementById("CN_txtGioiTinh").GetValue()?.ToLower())
                 {
                     case "nam":
                         accInfo.Gender = Gender.Male;
@@ -311,24 +333,30 @@ namespace DutWrapper
                 }
 
                 accInfo.NationalCardID = document.GetElementById("CN_txtSoCMND").GetValue();
-                accInfo.NationalCardIssueDate = DateTime.ParseExact(document.GetElementById("CN_txtNgayCap").GetValue(), "dd/MM/yyyy", CultureInfo.InvariantCulture);
-                accInfo.NationalCardIssuePlace = document.GetElementById("CN_cboNoiCap").GetSelectedOptionOnSelectTag().TextContent;
+                accInfo.NationalCardIssueDate = document.GetElementById("CN_txtNgayCap").ConvertToDateTime();
+                accInfo.NationalCardIssuePlace = document.GetElementById("CN_cboNoiCap").GetSelectedOptionOnSelectTag().GetTextContent();
                 accInfo.CitizenCardID = document.GetElementById("CN_txtSoCCCD").GetValue();
-                accInfo.CitizenCardIssueDate = DateTime.ParseExact(document.GetElementById("CN_txtNcCCCD").GetValue(), "dd/MM/yyyy", CultureInfo.InvariantCulture);
+                accInfo.CitizenCardIssueDate = document.GetElementById("CN_txtNcCCCD").ConvertToDateTime();
+
+                accInfo.HealthInsuranceID = document.GetElementById("CN_txtSoBHYT").GetValue();
+                accInfo.HealthInsuranceExpirationDate = document.GetElementById("CN_txtHanBHYT").ConvertToDateTime();
 
                 accInfo.ClassName = document.GetElementById("CN_txtLop").GetValue();
+                accInfo.Specialization = document.GetElementById("MainContent_CN_txtNganh").GetValue();
+                accInfo.TrainingProgramPlan = document.GetElementById("MainContent_CN_txtCTDT").GetValue();
+                accInfo.TrainingProgramPlan2 = document.GetElementById("MainContent_CN_txtCT2").GetValue();
+
+                accInfo.SchoolEmail = document.GetElementById("CN_txtMail1").GetValue();
+                accInfo.PersonalEmail = document.GetElementById("CN_txtMail2").GetValue();
+                accInfo.FacebookLink = document.GetElementById("CN_txtFace").GetValue();
+                accInfo.PhoneNumber = document.GetElementById("CN_txtPhone").GetValue();
+
                 accInfo.BankInfo =
-                    String.Format(
+                    string.Format(
                         "{0} ({1})",
                         document.GetElementById("CN_txtTKNHang").GetValue(),
                         document.GetElementById("CN_txtNgHang").GetValue()
                         );
-                accInfo.HealthInsuranceID = document.GetElementById("CN_txtSoBHYT").GetValue();
-                accInfo.HealthInsuranceExpirationDate = DateTime.ParseExact(document.GetElementById("CN_txtHanBHYT").GetValue(), "dd/MM/yyyy", CultureInfo.InvariantCulture);
-                accInfo.EducationEmail = document.GetElementById("CN_txtMail1").GetValue();
-                accInfo.PersonalEmail = document.GetElementById("CN_txtMail2").GetValue();
-                accInfo.FacebookLink = document.GetElementById("CN_txtFace").GetValue();
-                accInfo.PhoneNumber = document.GetElementById("CN_txtPhone").GetValue();
             }
             catch
             {
@@ -336,6 +364,103 @@ namespace DutWrapper
             }
 
             return accInfo;
+        }
+
+        public static async Task<AccountTrainingResult?> GetAccountTrainingResult(string sessionId)
+        {
+            AccountTrainingResult? result = null;
+
+            try
+            {
+                var client = CreateDefaultHttpClient(sessionId);
+                HttpResponseMessage response = await client.GetAsync($"/PageKQRL.aspx");
+
+                if (!response.IsSuccessStatusCode)
+                    throw new Exception(String.Format("The request has return code {0}.", response.StatusCode));
+
+                var config = Configuration.Default;
+                var context = BrowsingContext.New(config);
+                var document = await context.OpenAsync(req => req.Content(response.Content.ReadAsStringAsync().Result));
+
+                TrainingSummary trainSum = new TrainingSummary();
+                var docTrainList = document.GetElementById("KQRLGridTH").GetElementsByClassName("GridRow");
+                foreach (var item in docTrainList)
+                {
+                    var cellList = item.GetElementsByClassName("GridCell");
+                    if (cellList[0].GetTextContent().IsNullOrEmpty() ||
+                        cellList[cellList.Length - 1].GetTextContent().IsNullOrEmpty() ||
+                        cellList[cellList.Length - 2].GetTextContent().IsNullOrEmpty() ||
+                        cellList[cellList.Length - 3].GetTextContent().IsNullOrEmpty()
+                        ) continue;
+
+                    if (trainSum.SchoolYearStart.IsNullOrEmpty())
+                    {
+                        trainSum.SchoolYearStart = cellList[0].GetTextContent();
+                    }
+
+                    trainSum.SchoolYearCurrent = cellList[0].GetTextContent();
+                    trainSum.CreditCollected = Convert.ToDouble(cellList[cellList.Length - 3].GetTextContent());
+                    trainSum.AvgTrainingScore4 = Convert.ToDouble(cellList[cellList.Length - 2].GetTextContent());
+                    trainSum.AvgSocial = Convert.ToDouble(cellList[cellList.Length - 1].GetTextContent());
+
+                }
+
+                GraduateSummary gradSum = new GraduateSummary();
+                var docGrad = document.GetElementById("KQRLdvCc").ConvertToIDocument();
+                gradSum.HasSigGDTC = docGrad.GetElementById("KQRL_chkGDTC").IsSelectedInInput();
+                gradSum.HasSigGDQP = docGrad.GetElementById("KQRL_chkQP").IsSelectedInInput();
+                gradSum.HasSigEnglish = docGrad.GetElementById("KQRL_chkCCNN").IsSelectedInInput();
+                gradSum.HasSigIT = docGrad.GetElementById("KQRL_chkCCTH").IsSelectedInInput();
+                gradSum.HasQualifiedGraduate = docGrad.GetElementById("KQRL_chkCNTN").IsSelectedInInput();
+                gradSum.Info1 = docGrad.GetElementById("KQRL_txtKT").GetTextContent();
+                gradSum.Info2 = docGrad.GetElementById("KQRL_txtKL").GetTextContent();
+                gradSum.Info3 = docGrad.GetElementById("KQRL_txtInfo").GetTextContent();
+                gradSum.ApproveGraduateProcessInfo = docGrad.GetElementById("KQRL_txtCNTN").GetTextContent();
+
+                List<SubjectResult> subSum = new List<SubjectResult>();
+                var docSub = document.GetElementById("KQRL_divContent")
+                    .ConvertToIDocument()
+                    .GetElementById("KQRLGridKQHT")
+                    .GetElementsByClassName("GridRow");
+                for (int i = docSub.Length - 1; i >= 0; i--)
+                {
+                    var docCell = docSub[i].GetElementsByClassName("GridCell");
+
+                    var item = new SubjectResult();
+                    item.SchoolYear = docCell[1].GetTextContent();
+                    item.IsExtendedSemester = docCell[2].ClassList.Contains("GridCheck");
+                    item.ID = docCell[3].GetTextContent();
+                    item.Name = docCell[4].GetTextContent();
+                    item.Credit = double.Parse(docCell[5].GetTextContent());
+                    item.PointFormula = docCell[6].GetTextContent();
+                    item.PointBT = docCell[7].GetTextContent().IsNullOrEmpty() ? new double?() : double.Parse(docCell[7].GetTextContent());
+                    item.PointBV = docCell[8].GetTextContent().IsNullOrEmpty() ? new double?() : double.Parse(docCell[8].GetTextContent());
+                    item.PointCC = docCell[9].GetTextContent().IsNullOrEmpty() ? new double?() : double.Parse(docCell[9].GetTextContent());
+                    item.PointCK = docCell[10].GetTextContent().IsNullOrEmpty() ? new double?() : double.Parse(docCell[10].GetTextContent());
+                    item.PointGK = docCell[11].GetTextContent().IsNullOrEmpty() ? new double?() : double.Parse(docCell[11].GetTextContent());
+                    item.PointQT = docCell[12].GetTextContent().IsNullOrEmpty() ? new double?() : double.Parse(docCell[12].GetTextContent());
+                    item.PointTH = docCell[13].GetTextContent().IsNullOrEmpty() ? new double?() : double.Parse(docCell[13].GetTextContent());
+                    item.PointFinalT10 = docCell[14].GetTextContent().IsNullOrEmpty() ? new double?() : double.Parse(docCell[14].GetTextContent());
+                    item.PointFinalT4 = docCell[15].GetTextContent().IsNullOrEmpty() ? new double?() : double.Parse(docCell[15].GetTextContent());
+                    item.PointFinalByChar = docCell[16].GetTextContent() ?? "I";
+                    item.IsReStudy = subSum.Any(p => p.Name.ToLower() == docCell[4].GetTextContent().ToLower());
+
+                    subSum.Add(item);
+                }
+
+                result = new AccountTrainingResult()
+                {
+                    TrainingSummary = trainSum,
+                    GraduateSummary = gradSum,
+                    SubjectResultList = subSum
+                };
+            }
+            catch
+            {
+
+            }
+
+            return result;
         }
 
         private static T ConvertTo<T>(string str)
@@ -367,16 +492,6 @@ namespace DutWrapper
             }
 
             return result;
-        }
-
-        private static string? GetValue(this IElement element)
-        {
-            return element.GetAttribute("value");
-        }
-
-        private static IElement? GetSelectedOptionOnSelectTag(this IElement element)
-        {
-            return element.GetElementsByTagName("option").ToList().FirstOrDefault(p => p.HasAttribute("selected"));
         }
     }
 }