+++ DutWrapper/Model/Enums/Gender.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace DutWrapper.Model.Enums
+{
+    public enum Gender
+    {
+        Unknown = -1,
+        Male,
+        Female,
+    }
+}